+++ src/app/_components/review/review-component/IconChipList.tsx
@@ -1,5 +1,5 @@
 import { IconChipListProps } from '@/types/map/ReviewProps';
-import IconContainer from '@/app/service/map/_components/review-component/IconContainer';
+import IconContainer from '@/app/_components/review/review-component/IconContainer';
 import { Chip, Stack } from '@mui/material';
 import { v4 } from 'uuid';
 import { useMemo } from 'react';
