+++ src/app/_components/review/review-component/SatisfiedRating.tsx
@@ -7,8 +7,8 @@ import { RadioGroupRatingProps } from '@/types/map/ReviewProps';
 import { useEffect } from 'react';
 import IconContainer, {
   customIcons,
-} from '@/app/service/map/_components/review-component/IconContainer';
-import { IconChip } from '@/app/service/map/_components/review-component/IconChipList';
+} from '@/app/_components/review/review-component/IconContainer';
+import { IconChip } from '@/app/_components/review/review-component/IconChipList';
 
 const SatisfiedRating = styled(Rating)(({ theme }) => ({
   '& .MuiRating-iconEmpty .MuiSvgIcon-root': {
